Dot notation for partials.
A weakness of previous versions was an inability to distinguish between a dot character in a partial and a class name. An update to the regex that initially splits the line fixes this along with minor tweaks.

Also included a third param to the plugin so that the partial and variable lookup objects can be passed in separately. This was not added to the docs for now and will be considered a hidden feature until I can put the concept through some paces. Should work well for Backbone.js view integration.